Initial push, setting up classes, interfaces and scriptable classes for attackers and defenders
