--- Assets/Cells.cs
@@ -1,4 +1,0 @@
-public class Cells
-{
-    public bool Iswater;
-}
