+++ Assets/Scripts/Characters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 06734f599aeda18439bd8c4716747a09
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
