+++ Assets/Scripts/Characters/Attackers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cfcbdb63b6159e144b03110e099cb399
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
