+++ Assets/Scripts/Characters/Attackers/EnemyScriptable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0396f644d23f5d246b5eb5b6efd150f7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
