+++ Assets/Scripts/Characters/Attackers/EnemySuper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8b0fdcbfd5684c849a9e2a51ce5f4ffd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
