+++ Assets/Scripts/Characters/Attackers/IEnemy.cs
@@ -1,0 +1,11 @@
+namespace Characters.Attackers
+{
+   public interface IEnemy 
+   {
+   
+      void Attack();
+      void Move();
+      void TakeDamage();
+      void Die();
+   }
+}
