+++ Assets/Scripts/Characters/Attackers/IEnemy.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b13de4ab8a3b38a4d88d30a69620a7a0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
