+++ Assets/Scripts/Characters/Attackers/MeleeEnemy.cs
@@ -1,0 +1,38 @@
+namespace Characters.Attackers
+{
+    public class MeleeEnemy : EnemySuper, IEnemy
+    {
+    
+        // Start is called before the first frame update
+        void Start()
+        {
+        
+        }
+
+        public void Attack()
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public void Move()
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public void TakeDamage()
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public void Die()
+        {
+            throw new System.NotImplementedException();
+        }
+
+        // Update is called once per frame
+        void Update()
+        {
+        
+        }
+    }
+}
