+++ Assets/Scripts/Characters/Defenders.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 36b3ae7ad5e3a004281bd4e68c2c22da
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
