+++ Assets/Scripts/Characters/Defenders/DefenderScriptable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3a2695900db7eba4996e21aaec5a9aab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
