+++ Assets/Scripts/Characters/Defenders/DefenderSuper.cs
@@ -1,0 +1,22 @@
+using UnityEngine;
+
+namespace Characters.Defenders
+{
+    public class DefenderSuper : MonoBehaviour
+    {
+        public GameObject target;
+        public ScriptableObject defenderScript;
+        public float maxHealth, damage, moveSpeed;
+        // Start is called before the first frame update
+        void Start()
+        {
+        
+        }
+
+        // Update is called once per frame
+        void Update()
+        {
+        
+        }
+    }
+}
