+++ Assets/Scripts/Characters/Defenders/DefenderSuper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bb71917c2e57ea5428715d1874025629
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
