+++ Assets/Scripts/Characters/Defenders/IDefender.cs
@@ -1,0 +1,10 @@
+namespace Characters.Defenders
+{
+    public interface IDefender
+    {
+        void Move();
+        void Attack();
+        void TakeDamage();
+        void Death();
+    }
+}
