+++ Assets/Scripts/Characters/Defenders/IDefender.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 23477d556a9900a4dad67b0bd54a57e0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
