+++ Assets/Scripts/Characters/Defenders/MeleeDefender.cs
@@ -1,0 +1,38 @@
+namespace Characters.Defenders
+{
+    public class MeleeDefender : DefenderSuper, IDefender
+    {
+        // Start is called before the first frame update
+        void Start()
+        {
+        
+        }
+    
+        public void Move()
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public void Attack()
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public void TakeDamage()
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public void Death()
+        {
+            throw new System.NotImplementedException();
+        }
+        // Update is called once per frame
+        void Update()
+        {
+        
+        }
+
+    
+    }
+}
