+++ Assets/Scripts/Characters/Defenders/MeleeDefender.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3d4c6fac0f3fdea4f8c8c949532f2210
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
