+++ Assets/Scripts/Characters/Defenders/Test.cs
@@ -1,0 +1,10 @@
+
+using UnityEngine;
+
+namespace Characters.Defenders
+{
+    public class Test : MonoBehaviour
+    {
+    
+    }
+}
