+++ Assets/Scripts/Characters/Defenders/Test.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1ec54462164e29c49a661a2989166da0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
