+++ Assets/Scripts/Terrain/Cells.cs
@@ -1,0 +1,7 @@
+namespace Terrain
+{
+    public class Cells
+    {
+        public bool Iswater;
+    }
+}
